Updated compile-time parsing of a bignum-literal
The topic related to this commit is the conversion of a string of
digits representing a base-10 encoding of a big-integer
into a base-2^w representation of the same integer, where w is
the machine wordsize (typically, w=64).

Previously, due to implementation details, the compiler performed
this conversion twice for every use of the _Z user-defined literal,
which is wasteful (in terms of compilation time).
This commit removes this redundancy; now only one conversion is performed
for every use of the _Z suffix.